pyproject.toml
[project]
name = "demo-app-service"
version = "0.1.0"
description = ""
authors = [
    {name = "mjthecoder65",email = "michael53161@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic (>=1.17.2,<2.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "fastapi[standard] (>=0.124.4,<0.125.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "google-cloud-storage (==2.14.0)",
    "google-cloud-firestore (>=2.21.0,<3.0.0)",
    "google-generativeai (==0.8.3)",
    "pillow (>=12.0.0,<13.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]

[tool.pytest.ini_options]
pythonpath = "."